Add `assign_to_tuple`
